# backend/docs/doc_service_openapi.json
{"openapi":"3.1.0","info":{"title":"Document Processing Service","description":"Microservice for document processing with GOST formatting","version":"1.0.0"},"paths":{"/documents/extract":{"post":{"tags":["documents"],"summary":"Extract","operationId":"extract_documents_extract_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_extract_documents_extract_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/health":{"get":{"tags":["documents"],"summary":"Health","operationId":"health_documents_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/documents/apply_ml_changes":{"post":{"tags":["documents"],"summary":"Apply Ml Changes Endpoint","description":"ÐÐ¾Ð»Ð½ÑÐ¹ Ð¿Ð°Ð¹Ð¿Ð»Ð°Ð¹Ð½: Ð¸Ð·Ð²Ð»ÐµÑÐµÐ½Ð¸Ðµ â ML Ð¿ÑÐ°Ð²ÐºÐ¸ â ÑÐ±Ð¾ÑÐºÐ° â ÐÐÐ¡Ð¢ â ÑÐ¸ÑÑÐ»ÑÐ½ÑÐ¹ Ð»Ð¸ÑÑ\n\nÐÑÐ¸Ð½Ð¸Ð¼Ð°ÐµÑ JSON:\n{\n    \"project_id\": \"string\",\n    \"ml_response\": { ... },  # Ð¿Ð¾Ð»Ð½ÑÐ¹ Ð¾ÑÐ²ÐµÑ Ð¾Ñ ML\n    \"title_page\": { ... }     # Ð´Ð°Ð½Ð½ÑÐµ ÑÐ¸ÑÑÐ»ÑÐ½Ð¾Ð³Ð¾ Ð»Ð¸ÑÑÐ°\n}\n\nÐÐ¾Ð·Ð²ÑÐ°ÑÐ°ÐµÑ Ð³Ð¾ÑÐ¾Ð²ÑÐ¹ DOCX ÑÐ°Ð¹Ð»","operationId":"apply_ml_changes_endpoint_documents_apply_ml_changes_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/documents/download/{project_id}":{"get":{"tags":["documents"],"summary":"Download Result","description":"Ð¡ÐºÐ°ÑÐ°ÑÑ Ð³Ð¾ÑÐ¾Ð²ÑÐ¹ Ð´Ð¾ÐºÑÐ¼ÐµÐ½Ñ Ð¿Ð¾ project_id\n\nÐÑÐµÑ ÑÐ°Ð¹Ð»Ñ Ð² Ð¿Ð¾ÑÑÐ´ÐºÐµ Ð¿ÑÐ¸Ð¾ÑÐ¸ÑÐµÑÐ°:\n1. {project_id}_final.docx\n2. {project_id}_result.docx  \n3. {project_id}.docx","operationId":"download_result_documents_download__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/info/{project_id}":{"get":{"tags":["documents"],"summary":"Get Project Info","description":"ÐÐ¾Ð»ÑÑÐ¸ÑÑ Ð¸Ð½ÑÐ¾ÑÐ¼Ð°ÑÐ¸Ñ Ð¾ Ð¿ÑÐ¾ÐµÐºÑÐµ: ÐºÐ°ÐºÐ¸Ðµ ÑÐ°Ð¹Ð»Ñ ÑÑÑÐµÑÑÐ²ÑÑÑ","operationId":"get_project_info_documents_info__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/download-pdf/{project_id}":{"get":{"tags":["documents"],"summary":"Download Pdf","description":"Ð¡ÐºÐ°ÑÐ°ÑÑ Ð³Ð¾ÑÐ¾Ð²ÑÐ¹ Ð´Ð¾ÐºÑÐ¼ÐµÐ½Ñ Ð² ÑÐ¾ÑÐ¼Ð°ÑÐµ PDF\n\nÐÐ¾Ð½Ð²ÐµÑÑÐ¸ÑÑÐµÑ DOCX Ð² PDF Ð¸ Ð²Ð¾Ð·Ð²ÑÐ°ÑÐ°ÐµÑ ÑÐ°Ð¹Ð».\nÐÑÐ»Ð¸ PDF ÑÐ¶Ðµ ÑÑÑÐµÑÑÐ²ÑÐµÑ, Ð²Ð¾Ð·Ð²ÑÐ°ÑÐ°ÐµÑ ÐµÐ³Ð¾.","operationId":"download_pdf_documents_download_pdf__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Healthcheck","operationId":"healthcheck__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_extract_documents_extract_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"project_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Id"}},"type":"object","required":["file"],"title":"Body_extract_documents_extract_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}